requirements.txt
backcall==0.1.0
decorator==4.4.2
entrypoints==0.3
flake8==3.7.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
mccabe==0.6.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyYAML==5.3
six==1.14.0
traitlets==4.3.3
wcwidth==0.1.8
